cleanup: move NodeSchedulingPropertiesChange
